+++ protected/humhub/modules/user/messages/cs/views_auth_recoverPassword.php
@@ -2,9 +2,9 @@
 return array (
   '<strong>Password</strong> recovery' => '<strong>Obnovení</strong> hesla',
   'Back' => 'Zpět',
-  'Just enter your e-mail address. We´ll send you recovery instructions!' => 'Stačí zadat vaši e-mailovou adresu, na ni vám zašleme instrukce k obnovení hesla.',
+  'Just enter your e-mail address. We\'ll send you recovery instructions!' => 'Stačí zadat vaši e-mailovou adresu, na ni vám zašleme instrukce k obnovení hesla.',
   'Password recovery' => 'Obnovení hesla',
   'Reset password' => 'Obnovit heslo',
-  'enter security code above' => 'Zadejte zabezpečovací kód',
-  'your email' => 'Vaše e-mailová adresa',
+  'Enter security code above' => 'Zadejte zabezpečovací kód',
+  'Your email' => 'Vaše e-mailová adresa',
 );