+++ protected/humhub/modules/user/messages/da/views_auth_recoverPassword.php
@@ -5,6 +5,6 @@ return array (
   'Just enter your e-mail address. We´ll send you recovery instructions!' => 'Bare indtast din e-mail adresse. Vi sender dig nulstillings instruktioner!',
   'Password recovery' => 'Genoprettelse af adgangskode',
   'Reset password' => 'Nulstil adgangskode',
-  'enter security code above' => 'indtast sikkerheds kode ovenover',
-  'your email' => 'din e-mail',
+  'Enter security code above' => 'indtast sikkerheds kode ovenover',
+  'Your email' => 'din e-mail',
 );