+++ protected/humhub/modules/user/messages/id/views_auth_recoverPassword.php
@@ -2,9 +2,9 @@
 return array (
   '<strong>Password</strong> recovery' => '',
   'Back' => 'Kembali',
-  'Just enter your e-mail address. We´ll send you recovery instructions!' => '',
+  'Just enter your e-mail address. We\'ll send you recovery instructions!' => '',
   'Password recovery' => '',
   'Reset password' => '',
-  'enter security code above' => '',
-  'your email' => '',
+  'Enter security code above' => '',
+  'Your email' => '',
 );