+++ protected/humhub/modules/user/messages/ja/views_auth_recoverPassword.php
@@ -18,10 +18,10 @@
  */
 return [
     '<strong>Password</strong> recovery' => '',
-    'Just enter your e-mail address. We´ll send you recovery instructions!' => '',
+    'Just enter your e-mail address. We\'ll send you recovery instructions!' => '',
     'Password recovery' => '',
     'Reset password' => '',
-    'enter security code above' => '',
-    'your email' => '',
+    'Enter security code above' => '',
+    'Your email' => '',
     'Back' => '戻る',
 ];