+++ protected/humhub/modules/user/messages/nb_no/views_auth_recoverPassword.php
@@ -2,9 +2,9 @@
 return array (
   '<strong>Password</strong> recovery' => '<strong>Gjenopprett</strong> passord',
   'Back' => 'Tilbake',
-  'Just enter your e-mail address. We´ll send you recovery instructions!' => 'Tast inn din e-post adresse, vi sender deg informasjon om hvordan du gjenoppretter passordet ditt!',
+  'Just enter your e-mail address. We\'ll send you recovery instructions!' => 'Tast inn din e-post adresse, vi sender deg informasjon om hvordan du gjenoppretter passordet ditt!',
   'Password recovery' => 'Gjenopprett passord',
   'Reset password' => 'Tilbakestill passord',
-  'enter security code above' => 'tast inn sikkerhetskoden over',
-  'your email' => 'din e-post adresse',
+  'Enter security code above' => 'tast inn sikkerhetskoden over',
+  'Your email' => 'din e-post adresse',
 );