+++ protected/humhub/modules/user/messages/tr/views_auth_recoverPassword.php
@@ -2,9 +2,9 @@
 return array (
   '<strong>Password</strong> recovery' => '<strong> Şifre </strong> kurtarma',
   'Back' => 'Geri',
-  'Just enter your e-mail address. We´ll send you recovery instructions!' => 'Sadece mail adresini gir. Sana şifreni kurtarma adımlarını göndereceğiz.',
+  'Just enter your e-mail address. We\'ll send you recovery instructions!' => 'Sadece mail adresini gir. Sana şifreni kurtarma adımlarını göndereceğiz.',
   'Password recovery' => 'Şifre kurtarma',
   'Reset password' => 'Şifre sıfırla',
-  'enter security code above' => 'Yeni Şifre',
-  'your email' => 'E-posta Adresin',
+  'Enter security code above' => 'Yeni Şifre',
+  'Your email' => 'E-posta Adresin',
 );